requirements.txt
absl-py==2.2.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
coverage==7.9.1
cron-descriptor==1.4.5
croniter==6.0.0
cssselect2==0.8.0
decorator==5.2.1
deepface==0.0.93
deprecation==2.1.0
diff-match-patch==20241021
dj-database-url==3.0.0
Django==5.2.1
django-appconf==1.1.0
django-celery-beat==2.8.0
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-email-verification==0.3.3
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-graphql-jwt==0.4.0
django-import-export==4.3.7
django-redis==5.4.0
django-select2==8.4.0
django-timezone-field==7.1
django-weasyprint==2.4.0
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.57.0
-e git+https://github.com/frappe/frappe-client@aad36de5eb27be1bc6a8357791ee26ecb7cecd99#egg=frappeclient
gast==0.6.0
gdown==5.2.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphene-gis==0.0.8
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
httplib2==0.22.0
icecream==2.1.4
idna==3.10
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
kombu==5.5.3
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.0.9
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.5
pillow==11.2.1
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
redis==5.3.0
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rsa==4.9.1
setuptools==80.4.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
text-unidecode==1.3
tf_keras==2.19.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
validators==0.35.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
XlsxWriter==3.2.3
zopfli==0.2.3.post1
